+++ tests/integration/test_sepa_direct_debit_account.py
@@ -18,6 +18,29 @@ class TestSepaDirectDebitAccount(unittest.TestCase):
         self.assertEqual(found_account.token, result.payment_method.token)
         self.assertTrue(found_account.global_id)
 
+    def test_find_returns_subscriptions_associated_with_a_sepa_direct_debit_account(self):
+        result = PaymentMethod.create({
+            "customer_id": Customer.create().customer.id,
+            "payment_method_nonce": Nonces.SepaDirectDebit
+        })
+        self.assertTrue(result.is_success)
+
+        token = result.payment_method.token
+
+        subscription1 = Subscription.create({
+            "payment_method_token": token,
+            "plan_id": TestHelper.trialless_plan["id"]
+        }).subscription
+
+        subscription2 = Subscription.create({
+            "payment_method_token": token,
+            "plan_id": TestHelper.trialless_plan["id"]
+        }).subscription
+
+        sepa_direct_debit_account = SepaDirectDebitAccount.find(result.payment_method.token)
+        self.assertTrue(subscription1.id in [s.id for s in sepa_direct_debit_account.subscriptions])
+        self.assertTrue(subscription2.id in [s.id for s in sepa_direct_debit_account.subscriptions])
+
     def test_find_raises_on_not_found_token(self):
         self.assertRaises(NotFoundError, SepaDirectDebitAccount.find, "non-existant-token")